// plugins/cad/src/types/ConfigManagement.ts
/**
 * Copyright 2022 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { KubernetesKeyValueObject } from './KubernetesResource';

export type ListConfigManagements = {
  kind: string;
  apiVersion: string;
  items: ConfigManagement[];
};

export type ConfigManagement = {
  apiVersion: string;
  kind: string;
  metadata: ConfigManagementMetadata;
  spec: ConfigManagementSpec;
  status?: ConfigManagementStatus;
};

export type ConfigManagementMetadata = {
  name: string;
  labels?: KubernetesKeyValueObject;
  annotations?: KubernetesKeyValueObject;
};

export type ConfigManagementSpec = {
  enableMultiRepo: boolean;
};

export type ConfigManagementStatus = {
  configManagementVersion?: string;
  healthy: boolean;
};